+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     * BUGFIX      #738 Removed action icon in datagrid if no icon is defined
     * ENHANCEMENT #735 Added maxHeight options to toolbar dropdownOptions
     * BUGFIX      #722 Fixed create hover button for ghost pages and made it optional
+    * FEATURE     #733 Created addImage functionality in dropzone
     * ENHANCEMENT #723 Added description class to itembox and made content take up the whole width
     * FEATURE     #714 Tabs and language Changer are now in a new section -> language Changer can be used without Tabs.
     * BUGFIX      #729 Update parent checkbox style when removing child in table-view datagrid-decorator